\name{amabie-package}
\alias{amabie-package}
\alias{amabie}
\docType{package}
\title{
\packageTitle{amabie}
}
\description{
\packageDescription{amabie}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{amabie}
\packageIndices{amabie}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{amabie}

Maintainer: \packageMaintainer{amabie}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
