\name{amabie-package}
\alias{amabie-package}
\alias{amabie}
\docType{package}
\title{
test
}
\description{
test
}
\details{

The DESCRIPTION file:
test
test
test
}
\author{
test
test
}
\references{
test
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
test
}
\examples{
test
}
